requirements/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain>=0.1.10,<0.2.0
langchain-openai>=0.0.2,<0.1.0
openai>=1.6.1,<2.0.0
faiss-cpu==1.7.4
numpy>=1.22.0,<1.27.0
pypdf==3.17.0
tiktoken>=0.5.2,<0.6.0
crewai==0.28.0
PyPDF2>=3.0.0
slowapi==0.1.8
redis==5.0.1
prometheus-client==0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic>=0.4.27
aiofiles>=23.2.1
streamlit>=1.32.0
requests>=2.31.0
python-docx>=0.8.11
